docs(di-cookbook): InjectionToken->OpaqueToken; clarify class-interface
