+++ public/docs/_examples/cb-dependency-injection/ts/src/app/date-logger.service.ts
@@ -4,29 +4,10 @@ import { Injectable }    from '@angular/core';
 
 import { LoggerService } from './logger.service';
 
-// #docregion minimal-logger
-// class used as a restricting interface (hides other public members)
-export abstract class MinimalLogger {
-  logInfo: (msg: string) => void;
-  logs: string[];
-}
-// #enddocregion minimal-logger
-
-/*
-// Transpiles to:
-// #docregion minimal-logger-transpiled
-  var MinimalLogger = (function () {
-    function MinimalLogger() {}
-    return MinimalLogger;
-  }());
-  exports("MinimalLogger", MinimalLogger);
-// #enddocregion minimal-logger-transpiled
- */
-
 // #docregion date-logger-service
 @Injectable()
 // #docregion date-logger-service-signature
-export class DateLoggerService extends LoggerService implements MinimalLogger
+export class DateLoggerService extends LoggerService
 // #enddocregion date-logger-service-signature
 {
   logInfo(msg: any)  { super.logInfo(stamp(msg)); }
