+++ public/docs/_examples/cb-dependency-injection/ts/src/app/hero-of-the-month.component.1.ts
@@ -1,0 +1,26 @@
+// Illustrative (not used), mini-version of the actual HeroOfTheMonthComponent
+// Injecting with the MinimalLogger "interface-class"
+import { Component, NgModule } from '@angular/core';
+import { LoggerService }       from './logger.service';
+import { MinimalLogger }       from './minimal-logger.service';
+
+// #docregion
+@Component({
+  selector: 'hero-of-the-month',
+  templateUrl: './hero-of-the-month.component.html',
+  // Todo: move this aliasing, `useExisting` provider to the AppModule
+  providers: [{ provide: MinimalLogger, useExisting: LoggerService }]
+})
+export class HeroOfTheMonthComponent {
+  logs: string[] = [];
+  constructor(logger: MinimalLogger) {
+    logger.logInfo('starting up');
+  }
+}
+// #enddocregion
+
+// This NgModule exists only to avoid the Angular language service's "undeclared component" error
+@NgModule({
+  declarations: [ HeroOfTheMonthComponent ]
+})
+class NoopModule {}
