+++ public/docs/_examples/cb-dependency-injection/ts/src/app/hero-of-the-month.component.ts
@@ -1,19 +1,19 @@
 /* tslint:disable:one-line:check-open-brace*/
 // #docplaster
-// #docregion opaque-token
-import { OpaqueToken } from '@angular/core';
+// #docregion injection-token
+import { InjectionToken } from '@angular/core';
 
-export const TITLE = new OpaqueToken('title');
-// #enddocregion opaque-token
+export const TITLE = new InjectionToken<string>('title');
+// #enddocregion injection-token
 
 // #docregion hero-of-the-month
 import { Component, Inject } from '@angular/core';
 
-import { DateLoggerService,
-         MinimalLogger }     from './date-logger.service';
+import { DateLoggerService } from './date-logger.service';
 import { Hero }              from './hero';
 import { HeroService }       from './hero.service';
 import { LoggerService }     from './logger.service';
+import { MinimalLogger }     from './minimal-logger.service';
 import { RUNNERS_UP,
          runnersUpFactory }  from './runners-up';
 
@@ -22,28 +22,16 @@ import { RUNNERS_UP,
 const someHero = new Hero(42, 'Magma', 'Had a great month!', '555-555-5555');
 // #enddocregion some-hero
 
-const template = `
-  <h3>{{title}}</h3>
-  <div>Winner: <strong>{{heroOfTheMonth.name}}</strong></div>
-  <div>Reason for award: <strong>{{heroOfTheMonth.description}}</strong></div>
-  <div>Runners-up: <strong id="rups1">{{runnersUp}}</strong></div>
-
-  <p>Logs:</p>
-  <div id="logs">
-    <div *ngFor="let log of logs">{{log}}</div>
-  </div>
-  `;
-
 // #docregion hero-of-the-month
 @Component({
   selector: 'hero-of-the-month',
-  template: template,
+  templateUrl: './hero-of-the-month.component.html',
   providers: [
     // #docregion use-value
     { provide: Hero,          useValue:    someHero },
-    // #docregion provide-opaque-token
+    // #docregion provide-injection-token
     { provide: TITLE,         useValue:   'Hero of the Month' },
-    // #enddocregion provide-opaque-token
+    // #enddocregion provide-injection-token
     // #enddocregion use-value
     // #docregion use-class
     { provide: HeroService,   useClass:    HeroService },
@@ -52,9 +40,9 @@ const template = `
     // #docregion use-existing
     { provide: MinimalLogger, useExisting: LoggerService },
     // #enddocregion use-existing
-    // #docregion provide-opaque-token, use-factory
+    // #docregion provide-injection-token, use-factory
     { provide: RUNNERS_UP,    useFactory:  runnersUpFactory(2), deps: [Hero, HeroService] }
-    // #enddocregion provide-opaque-token, use-factory
+    // #enddocregion provide-injection-token, use-factory
   ]
 })
 export class HeroOfTheMonthComponent {
