+++ public/docs/_examples/cb-dependency-injection/ts/src/app/runners-up.ts
@@ -1,12 +1,12 @@
 // #docplaster
 // #docregion
-import { OpaqueToken } from '@angular/core';
+import { InjectionToken } from '@angular/core';
 
 import { Hero }        from './hero';
 import { HeroService } from './hero.service';
 
 // #docregion runners-up
-export const RUNNERS_UP = new OpaqueToken('RunnersUp');
+export const RUNNERS_UP = new InjectionToken<string>('RunnersUp');
 // #enddocregion runners-up
 
 // #docregion factory-synopsis
