+++ public/docs/ts/latest/guide/dependency-injection.jade
@@ -39,7 +39,7 @@ block includes
   * [Dependency injection tokens](#dependency-injection-tokens)
  
     * [Non-class dependencies](#non-class-dependencies)
-    * [`OpaqueToken`](#opaquetoken)
+    * [`InjectionToken`](#injection-token)
 
   * [Optional dependencies](#optional)
   * [Summary](#summary)
@@ -645,7 +645,7 @@ a#value-provider
 :marked
   See more `useValue` examples in the
   [Non-class dependencies](#non-class-dependencies) and
-  [OpaqueToken](#opaquetoken) sections.
+  [InjectionToken](#injection-token) sections.
 
 #factory-provider
 :marked
@@ -784,18 +784,20 @@ p
     The TypeScript interface disappears from the generated JavaScript.
     There is no interface type information left for Angular to find at runtime.
 
-a#opaquetoken
+a#injection-token
 :marked
-  ### _OpaqueToken_
+  ### _InjectionToken_
 
   One solution to choosing a provider token for non-class dependencies is
-  to define and use an <a href="../api/core/index/OpaqueToken-class.html"><b>OpaqueToken</b></a>.
-  The definition looks like this:
+  to define and use an <a href="../api/core/index/InjectionToken-class.html"><b>InjectionToken</b></a>.
+  The definition of such a token looks like this:
 
 +makeExample('dependency-injection/ts/src/app/app.config.ts','token')(format='.')
-
 :marked
-  Register the dependency provider using the `OpaqueToken` object:
+  The type parameter, while optional, conveys the dependency's type to developers and tooling.
+  The token description is another developer aid.
+
+  Register the dependency provider using the `InjectionToken` object:
 
 +makeExample('dependency-injection/ts/src/app/providers.component.ts','providers-9')(format=".")
